Move stress integration common method to transaction utils
